Added player interaction and money logic
Added a while loop to simulate the game.

Added user input validation

Added a money system.

The game still needs a proper rewards for winning and also all the different type of bets a user can choose in a sic bo game ( like only selecting two 4's or just big or small ).     The code is starting to get clunky......we could either make it more efficient or start making sub-functions.